requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.4
django-widget-tweaks==1.5.0
lxml==4.9.3
mysqlclient==2.2.0
Pillow==10.0.0
psycopg2==2.9.7
pytils==0.4.1
six==1.16.0
sqlparse==0.4.4
tools==0.1.9
typing_extensions==4.7.1
tzdata==2023.3
